+++ src/recommend.ts
@@ -13,7 +13,7 @@ import { capacitorDevicesCommand, capacitorRun } from './capacitor-run';
 import { capacitorRecommendations, checkCapacitorRules } from './rules-capacitor';
 import { checkCordovaPlugins, checkCordovaRules } from './rules-cordova';
 import { webProject } from './rules-web-project';
-import { checkPackages } from './rules-packages';
+import { checkPackages, checkRemoteDependencies } from './rules-packages';
 import { checkDeprecatedPlugins } from './rules-deprecated-plugins';
 import { capacitorSync } from './capacitor-sync';
 import { capacitorOpen } from './capacitor-open';
@@ -184,6 +184,8 @@ export async function getRecommendations(
       `${deprecated.name} is deprecated: ${deprecated.message}`
     );
   }
+
+  checkRemoteDependencies(project);
 
   // Deprecated plugins
   checkDeprecatedPlugins(project);